Reorganized styles into directory, fixed memoization in map, added useeffect to simulate loading markers, cleaned up marker code
