+++ App.js
@@ -1,7 +1,7 @@
 import { StatusBar } from "expo-status-bar";
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
-import styles from "./styles.js";
+import styles from "./styles/styles.js";
 import NavigationStack from "./navigation/NavigationStack";
 import Providers from "./navigation/index";
 import { LogBox } from "react-native";
